separate stubs logic from get_exceptions
