+++ deal/linter/_extractors/__init__.py
@@ -2,6 +2,7 @@
 from .common import get_name
 from .contracts import get_contracts
 from .exceptions import get_exceptions
+from .exceptions_stubs import get_exceptions_stubs
 from .globals import get_globals
 from .imports import get_imports
 from .prints import get_prints
@@ -11,6 +12,7 @@ from .returns import get_returns
 __all__ = [
     'get_contracts',
     'get_exceptions',
+    'get_exceptions_stubs',
     'get_globals',
     'get_imports',
     'get_name',
