+++ deal/linter/_extractors/exceptions.py
@@ -1,8 +1,7 @@
 # built-in
 import ast
 import builtins
-from pathlib import Path
-from typing import Iterator, Optional, Tuple
+from typing import Iterator
 
 # external
 import astroid
@@ -10,10 +9,9 @@ import astroid
 # app
 from .common import TOKENS, Token, get_name, traverse
 from .contracts import get_contracts
-from .._stub import StubsManager, StubFile, EXTENSION
 
 
-def get_exceptions(body: list, *, dive: bool = True, stubs: StubsManager = None) -> Iterator[Token]:
+def get_exceptions(body: list, *, dive: bool = True) -> Iterator[Token]:
     for expr in traverse(body):
         token_info = dict(line=expr.lineno, col=expr.col_offset)
 
@@ -67,9 +65,6 @@ def get_exceptions(body: list, *, dive: bool = True, stubs: StubsManager = None)
         if dive:
             yield from _exceptions_from_funcs(expr=expr)
 
-        if stubs is not None:
-            yield from _exceptions_from_stubs(expr=expr, stubs=stubs)
-
 
 def _exceptions_from_funcs(expr) -> Iterator[Token]:
     for name_node in get_names(expr):
@@ -108,75 +103,6 @@ def _exceptions_from_funcs(expr) -> Iterator[Token]:
                     if name is None:
                         continue
                     yield Token(value=name, **extra)
-
-
-def _exceptions_from_stubs(expr, stubs: StubsManager) -> Iterator[Token]:
-    if type(expr) is not astroid.Call:
-        return
-    try:
-        guesses = tuple(expr.func.infer())
-    except astroid.exceptions.InferenceError:
-        return
-    except RecursionError:
-        return
-    extra = dict(
-        line=expr.lineno,
-        col=expr.col_offset,
-    )
-    for value in guesses:
-        if type(value) is not astroid.FunctionDef:
-            continue
-        module_name, func_name = _get_full_name(expr=value)
-        stub = _get_stub(module_name=module_name, expr=value, stubs=stubs)
-        if stub is None:
-            continue
-        names = stub.get(func=func_name, contract='raises')
-        for name in names:
-            name = getattr(builtins, name, name)
-            yield Token(value=name, **extra)
-
-
-def _get_stub(module_name, expr, stubs: StubsManager) -> StubFile:
-    if not module_name:
-        return None
-    stub = stubs.get(module_name)
-    if stub is not None:
-        return stub
-
-    module = _get_module(expr=expr)
-    if module.file is None:
-        return None
-    path = Path(module.file).with_suffix(EXTENSION)
-    if not path.exists():
-        return None
-    return stubs.read(path=path)
-
-
-def _get_module(expr) -> Optional[astroid.Module]:
-    if type(expr) is astroid.Module:
-        return expr
-    if expr.parent is None:
-        return None
-    return _get_module(expr.parent)
-
-
-def _get_full_name(expr) -> Tuple[str, str]:
-    if expr.parent is None:
-        return '', expr.name
-
-    if type(expr.parent) is astroid.Module:
-        return expr.parent.qname(), expr.name
-
-    if type(expr.parent) is astroid.FunctionDef:
-        module_name, func_name = _get_full_name(expr=expr.parent)
-        return module_name, func_name + '.' + expr.name
-
-    if type(expr.parent) is astroid.ClassDef:
-        module_name, class_name = _get_full_name(expr=expr.parent)
-        return module_name, class_name + '.' + expr.name
-
-    path, _, func_name = expr.qname().rpartition('.')
-    return path, func_name
 
 
 def get_names(expr) -> Iterator:
