+++ deal/linter/_stub.py
@@ -1,4 +1,5 @@
 import json
+from itertools import chain
 from pathlib import Path
 from typing import Any, Dict, FrozenSet, Iterator, Optional, Sequence
 
@@ -129,7 +130,7 @@ def _get_funcs(*, path: Path) -> Iterator[PseudoFunc]:
 
 
 def generate_stub(*, path: Path, stubs: StubsManager = None) -> Path:
-    from ._extractors import get_exceptions
+    from ._extractors import get_exceptions, get_exceptions_stubs
 
     if path.suffix != '.py':
         raise ValueError('invalid Python file extension: *{}'.format(path.suffix))
@@ -138,7 +139,11 @@ def generate_stub(*, path: Path, stubs: StubsManager = None) -> Path:
         stubs = StubsManager()
     stub = stubs.create(path=path)
     for func in _get_funcs(path=path):
-        for token in get_exceptions(body=func.body, stubs=stubs):
+        tokens = chain(
+            get_exceptions(body=func.body),
+            get_exceptions_stubs(body=func.body, stubs=stubs),
+        )
+        for token in tokens:
             value = token.value
             if isinstance(value, type):
                 value = value.__name__
